requirements.txt
PyYAML==3.13
Scrapy==1.5.0
termcolor==1.1.0
